+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [2.11] - 2021-11-16
 ### Fixed
 - Fix some deprecated code causing warnings during automatic tests
+- Do not duplicate `mailto` schema when contact href is an email with correct schema
 ### Changed
 - Improve views code by reusing a controllers function when request auth fails
 ### Added